+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Airbrake Changelog
 
 * Read up to 4096 bytes from Rack request's body (increased from 512)
   ([#627](https://github.com/airbrake/airbrake/pull/627))
+* Fixed unwanted authentication when calling `current_user`, when Warden is
+  present ([#643](https://github.com/airbrake/airbrake/pull/643))
 
 ### [v5.6.1][v5.6.1] (October 24, 2016)